pyproject.toml
[project]
name = "ipv6-proxy-pool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "dnspython>=2.7.0",
    "pysocks>=1.7.1",
    "requests>=2.32.3",
    "socks>=0",
]
